[fix] 로그인 버튼 수정
로그인 후 새로고침 할 때 로그인 글씨가 잠깐 나타나는 현상 수정
